% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_get_avg_Fhat.R
\name{sup_get_avg_Fhat}
\alias{sup_get_avg_Fhat}
\title{Estimate average F.hat_j(t) for CDF pooling with sample splitting}
\usage{
sup_get_avg_Fhat(threshold, resid_fit)
}
\arguments{
\item{threshold}{Value t at which we compute average of F.hat_j(t)}

\item{resid_fit}{List of fitted residuals for supervised CDF pooling method}
}
\value{
Average F.hat_j(t) over the subjects in the residual fitting
group.
}
\description{
Given residuals R_ji = |Y_ji - mu.hat(X_ji)| and a threshold
t, compute F.hat_j(t) = (1/n_j) sum_{i=1}^{n_j} I(R_ji <= t).
Then return average F.hat_j(t) over the subjects in the residual fitting
group.
}
