% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_get_avg_pval.R
\name{sup_get_avg_pval}
\alias{sup_get_avg_pval}
\title{Function to get average p-val}
\usage{
sup_get_avg_pval(X_subsample_mat, Y_subsample_mat, X_new, Y_new)
}
\arguments{
\item{Y_new}{}
}
\value{

}
\description{
Function to get average p-val
}
