% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_repeated_subsample.R
\name{sup_repeated_subsample}
\alias{sup_repeated_subsample}
\title{Title}
\usage{
sup_repeated_subsample(
  xy_data,
  alpha,
  n_val,
  k_val,
  n_resamp,
  X_new,
  Y_new = NULL
)
}
\arguments{
\item{Y_new}{}
}
\value{

}
\description{
Title
}
