% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_single_subsample.R
\name{sup_single_subsample}
\alias{sup_single_subsample}
\title{Title}
\usage{
sup_single_subsample(xy_data, alpha, n_val, k_val, X_new, Y_new = NULL)
}
\arguments{
\item{Y_new}{}
}
\value{

}
\description{
Title
}
