% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsup_double_conformal.R
\name{unsup_double_conformal}
\alias{unsup_double_conformal}
\title{Construct unsupervised double conformal prediction interval.}
\usage{
unsup_double_conformal(Y, alpha, n_val, new_Y = NULL)
}
\arguments{
\item{n_val}{}
}
\value{

}
\description{
\enumerate{
\item Apply order statistic method to each of the k groups to get k intervals
\link{l_1, u_1}, ..., \link{l_k, u_k} each at level 1-alpha/2.
l_1 is the floor((n + 1) * alpha/4) order statistic of the first sample,
u_1 is the ceiling((n + 1)*(1 - alpha/4)) order statistic of the first sample,
and so on for the other samples.
\item Consider (l_1, ..., l_k). Take l_(r), where r = floor((k + 1) * alpha/4).
Consider (u_1, ..., u_k). Take u_(s), where s = ceiling((k + 1)*(1 - alpha/4)).
Prediction interval: C = \link{l_(r), u_(s)}.
}
}
