% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsup_get_avg_pvalue.R
\name{unsup_get_avg_pvalue}
\alias{unsup_get_avg_pvalue}
\title{Function to get average p-value}
\usage{
unsup_get_avg_pvalue(Y_subsample_mat, point)
}
\arguments{
\item{point}{}
}
\value{

}
\description{
Function to get average p-value
}
