% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsup_get_pval_shrinkage.R
\name{unsup_get_pval_shrinkage}
\alias{unsup_get_pval_shrinkage}
\title{Evaluate p-value of u using James-Stein shrinkage}
\usage{
unsup_get_pval_shrinkage(u, Y)
}
\arguments{
\item{u}{Hypothetical new observation on subject 1}

\item{Y}{List containing data of all subjects. Each item in the list
is a vector with one subject's observations.}
}
\value{
p-value of u relative to subject 1's sample, using the James-Stein
shrinkage estimate
}
\description{
Create augmented sample containing subject 1's observations
and u. Compute James-Stein shrinkage estimate of subject 1 mean using
augmented subject 1 sample and data from all other subjects.
As nonconformity scores, compute the absolute difference between each
observation in the augmented subject 1 sample and the James-Stein shrinkage
estimate. The p-value is the proportion of observations in the augmented
sample with nonconformity score >= u's nonconformity score.
}
