% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsup_method_group_1.R
\name{unsup_method_group_1}
\alias{unsup_method_group_1}
\title{Use only group 1's observations to construct prediction set for
a new group 1 observation}
\usage{
unsup_method_group_1(Y, alpha, Y_new = NULL)
}
\arguments{
\item{Y_new}{}
}
\value{

}
\description{
Use only group 1's observations to construct prediction set for
a new group 1 observation
}
