% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_pool_cdfs_split.R
\name{sup_pool_cdfs_split}
\alias{sup_pool_cdfs_split}
\title{Title}
\usage{
sup_pool_cdfs_split(
  xy_data,
  model_formula,
  alpha,
  k_val,
  k_model_fit,
  new_xy_data
)
}
\arguments{
\item{Y_new}{}
}
\value{

}
\description{
Title
}
