% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_repeated_subsample.R
\name{sup_repeated_subsample}
\alias{sup_repeated_subsample}
\title{Title}
\usage{
sup_repeated_subsample(
  xy_data,
  model_formula,
  alpha,
  n_val,
  k_indices,
  n_resamp,
  grid_values,
  new_xy_data,
  coverage_only = FALSE
)
}
\arguments{
\item{Y_new}{}
}
\value{

}
\description{
Title
}
