% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_single_subsample.R
\name{sup_single_subsample}
\alias{sup_single_subsample}
\title{Title}
\usage{
sup_single_subsample(
  xy_data,
  model_formula,
  alpha,
  n_val,
  k_indices,
  grid_values,
  new_xy_data
)
}
\arguments{
\item{Y_new}{}
}
\value{

}
\description{
Title
}
