% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsup_avg_ecdf.R
\name{unsup_avg_ecdf}
\alias{unsup_avg_ecdf}
\title{Evaluate average empirical CDF at a certain threshold}
\usage{
unsup_avg_ecdf(Y, threshold)
}
\arguments{
\item{threshold}{}
}
\value{

}
\description{
Evaluate average empirical CDF at a certain threshold
}
