% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsup_get_pval_group1.R
\name{unsup_get_pval_group1}
\alias{unsup_get_pval_group1}
\title{Evaluate p-value of potential y value Y_new compared to sample Y_1}
\usage{
unsup_get_pval_group1(Y_new, Y_1)
}
\arguments{
\item{Y_1}{}
}
\value{

}
\description{
Evaluate p-value of potential y value Y_new compared to sample Y_1
}
