% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsup_repeated_subsample.R
\name{unsup_repeated_subsample}
\alias{unsup_repeated_subsample}
\title{Title}
\usage{
unsup_repeated_subsample(Y, alpha, k_val, n_resamp, new_Y)
}
\arguments{
\item{new_Y}{(needs to be optional)}
}
\value{

}
\description{
Title
}
